Extract DescribeDescriptor
